[minishell] UPDATE: cd 리팩토링 중 (wip)
